[feat] 비디오 좋아요 컨트롤러 하나로 통합
좋아요 추가, 삭제 하나로 통합해서 deleteVideoLike 삭제 및 addLike 수정

+++ interview/src/main/java/inflearn/interview/service/VideoLikeService.java
@@ -21,19 +21,21 @@ public class VideoLikeService {
     private final VideoRepository videoRepository;
 
     public void addLike(Long video_id, User userDAO){
-        User user = userRepository.findById(userDAO.getUserId()).get();
-        Video video = videoRepository.findById(video_id).get();
-        VideoLike videoLike = new VideoLike();
-        videoLike.setVideo(video);
-        videoLike.setUser(user);
-        videoLike.setTime(LocalDateTime.now());
-
-        videoLikeRepository.save(videoLike);
-    }
-
-    public void deleteLike(Long video_id, User userDAO){
         Video video = videoRepository.findById(video_id).get();
         User user = userRepository.findById(userDAO.getUserId()).get();
-        videoLikeRepository.deleteByUserAndVideo(user, video);
+        VideoLike findResult = videoLikeRepository.findByUserAndVideo(user, video);
+        if(findResult == null){
+            VideoLike videoLike = new VideoLike();
+            videoLike.setVideo(video);
+            videoLike.setUser(user);
+            videoLike.setTime(LocalDateTime.now());
+
+            videoLikeRepository.save(videoLike);
+        }
+        else{
+            videoLikeRepository.deleteByUserAndVideo(user, video);
+        }
+
+
     }
 }
